图像处理/requirements.txt
matplotlib==2.2.2
numpy==1.14.2
opencv-python==3.4.1.15
Pillow==5.1.0
PyQt5==5.10.1
pyqt5-tools==5.9.0.1.2
pyqtgraph==0.10.0
scipy==1.1.0
sip==4.19.8
